* jc/revlist:
  rev-list --timestamp
  git-apply: do not barf when updating an originally empty file.
  http-push.c: squelch C90 warnings.
  fix field width/precision warnings in blame.c

+++ blame.c
@@ -748,7 +748,7 @@ int main(int argc, const char **argv)
 	struct commit_info ci;
 	const char *buf;
 	int max_digits;
-	size_t longest_file, longest_author;
+	int longest_file, longest_author;
 	int found_rename;
 
 	const char* prefix = setup_git_directory();
